fix: 커스텀 구분자로 "\n"이 올 수 있도록 수정
- DelimiterParser.parseCustomDelimiter()에서 커스텀 구분자를 추출할 때 문자열의 3번째 인덱스부터 "\n"(커스텀 구분자 접미어)를 찾도록 수정
- DelimiterParser.removeCustomDelimiterPrefix()에서 커스텀 구분자를 추출할 때 문자열의 3번째 인덱스부터 "\n"(커스텀 구분자 접미어)를 찾도록 수정

+++ src/main/java/calculator/Application.java
@@ -10,9 +10,9 @@ public class Application {
 
         // 계산
         StringSumCalculator stringSumCalculator = new StringSumCalculator();
-        double result = stringSumCalculator.calculate(input);
+        int result = stringSumCalculator.calculate(input);
 
         // 출력
-        System.out.printf("결과 : %f\n", result);
+        System.out.printf("결과 : %d\n", result);
     }
 }
